Move fixtures into a conftest
pytest will load them for us automatically and so we won't need the
flake8-upsetting imports from test_tahoe in test_streamedlogs.

also importing library code from test modules is sad.
